mcelog: a tool to access the machine check exceptions logged by the kernel.
The kernel indicated that machine check exceptions were logged, but did not tell what kind of exceptions.
In my case, this is usually the processor being throttled because it gets too hot (laptop...), but it
may also be more severe problems. Anyway, with this tool you can access these logs.


git-svn-id: https://nixos.org/repos/nix/nixpkgs/trunk@32547 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb